day07/requirements.txt
ollama
tavily-python
httpx
python-dotenv
